Crate roguelike

Structs

Game
The Game with accompanying state

Enums

Action
An Action represents something that happened, meta-data about the last move if you will. You could've picked something up, attacked or killed an enemy, walked into a wall or won. Nothing represents that you moved, but nothing happened, meaning you moved into an empty square.

Dir
Dir is an enum representing directions one could make a move in

Object
Object representing things on the map. An enemy needs to be hit twice in a row unless you have a sword, in which case they only need to be hit once in order to die. Gold is just extra points and takes no effort to pick up and reaching the goal finishes the game regardless of how many enemies you've killed.