Struct roguelike::Game

pub struct Game {
    pub position: (usize, usize),
    pub action: Action,
    // some fields omitted
}

The Game with accompanying state

Fields
position: (usize, usize)

Position of the player, defined as (row, column) coordinate on the world map

action: Action

Last action that you took

Methods

impl Game

fn new(print: bool) -> Game

Initialize a new game state and optionally print it

fn get_moves(&self) -> u32

Getter for the count of moves made

fn get_score(&self) -> u32

Getter for the current score

fn has_won(&self) -> bool

Getter for whether or not the game has been won

fn distance_to_goal(&self) -> f64

Euclidean distance to the goal

fn enter_move(&mut self, dir: &Dir, print: bool) -> bool

Enter move and optionally print the map, returning whether or not this move won the game

fn print_map(&self)

Prints the map on screen